Changed telephone and gsm field in user to accept characters instead of integers. Script now adds regions and rebuild database creates central and regional manager.

+++ rebuild_database.py
@@ -4,6 +4,7 @@ import getopt
 import django
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "wwinkel.settings")
 django.setup()
+from io import StringIO
 from django.core.management import call_command
 from django.utils import timezone
 from custom_users.models import User
@@ -11,9 +12,11 @@ from subprocess import call
 from glob import glob
 from cms.api import create_page
 from time import sleep
-
+from custom_users.models import ManagerUser, OrganisationUser, Region
 
 def main(argv):
+    os.chdir(os.path.dirname(os.path.realpath(__file__)))  # make sure working dir is dir where script is saved.
+
     solr_install_path = None
     try:
         opts, args = getopt.getopt(argv, "", ["solr_install_path="])
@@ -25,30 +28,65 @@ def main(argv):
             solr_install_path = arg
     # """
     if True:
-        print('deleting old db...')
-        try:
-            os.remove('db.sqlite3')
-        except FileNotFoundError:
-            print('no db.sqlite3 found')
+        print('testing need to delete the old db...')
+        should_delete = True
+        buf = StringIO()
+        call_command('makemigrations', stdout=buf)
+        buf.seek(0)
+        for line in buf.readlines():
+            if 'No changes detected' in line:
+                should_delete = False
+                break
+        print('  done...')
+
+        if should_delete:
+            print('deleting old db...')
+            try:
+                os.remove('db.sqlite3')
+            except FileNotFoundError:
+                print('no db.sqlite3 found')
+            print('  done')
+
+            print('deleting old migrations...')
+            migration_files = (
+                glob('./custom_users/migrations/*_auto_*.py') +
+                glob('./dbwwinkel/migrations/0001_initial.py') +
+                glob('./dbwwinkel/migrations/*_auto_*.py')
+            )
+            [os.remove(file) for file in migration_files]
+            print('  done')
+
+            print('generating new migrations...')
+            call_command('makemigrations', 'custom_users')
+            call_command('makemigrations')
+            print('  done')
+
+            print('executing migrations...')
+            call_command('migrate')
+            print('  done')
+
+    if solr_install_path:
+        os.chdir(solr_install_path)
+        temp_command = '.\\bin\\solr' if os.name == 'nt' else './bin/solr'
+        print('stopping solr if it is running...')
+        call([temp_command, 'stop', '-all'], shell=True)
+        sleep(5)
         print('  done')
 
-        print('deleting old migrations...')
-        migration_files = (
-            glob('./custom_users/migrations/*_auto_*.py') +
-            glob('./dbwwinkel/migrations/0001_initial.py') +
-            glob('./dbwwinkel/migrations/*_auto_*.py')
-        )
-        [os.remove(file) for file in migration_files]
+        print('generating haystack schema.xml...')
+        with open('example\solr\collection1\conf\schema.xml', mode='w') as f:
+            call_command('build_solr_schema', stdout=f)
         print('  done')
 
-        print('generating new migrations...')
-        call_command('makemigrations', 'custom_users')
-        call_command('makemigrations')
+        print('start solr...')
+        call([temp_command, 'start'], shell=True)
+        sleep(5)
         print('  done')
 
-        print('executing migrations...')
-        call_command('migrate')
+        print('rebuilding index...')
+        call_command('rebuild_index', '--noinput')
         print('  done')
+        os.chdir(os.path.dirname(os.path.realpath(__file__)))  # reset working directory.
 
     if True:
         print('importing data from csv...')
@@ -84,39 +122,47 @@ def main(argv):
         print('  done')
 
     if True:
-        print('creating superuser (user:admin@admin.be, pw:admin)...')
-        call_command('createsuperuser', '--email', 'admin@admin.be')
-        admin = User.objects.get(email='admin@admin.be')
-        admin.set_password('admin')
-        admin.save()
-        print('  done')
+        if not User.objects.filter(email='admin@admin.be').exists():
+            print('creating superuser (user:admin@admin.be, pw:admin)...')
+            call_command('createsuperuser', '--email', 'admin@admin.be')
+            admin = User.objects.get(email='admin@admin.be')
+            admin.set_password('admin')
+            admin.save()
+            print('  done')
+
+        if not User.objects.filter(email='central@manager.be').exists():
+            print('creating central manager (user:central@manager.be, pw:admin)...')
+            manager = ManagerUser.objects.update_or_create(
+                first_name='',
+                last_name='',
+                telephone='0987654321',
+                email='central@manager.be'
+            )
+            manager.region.add(Region.objects.get(id=Region.CENTRAL_REGION))
+            manager.set_password('admin')
+            manager.save()
+            print('  done')
+
+        if not User.objects.filter(email='regional@manager.be').exists():
+            print('creating regional manager (user:regional@manager.be, pw:admin)...')
+            manager = ManagerUser.objects.update_or_create(
+                first_name='',
+                last_name='',
+                telephone='0987654321',
+                email='regional@manager.be'
+            )
+            manager.region.add(Region.objects.get(id=Region.ANTWERP_REGION))
+            manager.set_password('admin')
+            manager.save()
+            print('  done')
+
+
     # """
     # print('creating manager user (user:central@manager.be, pw:admin)...')
 
 
     # print('creating regional user (user:regional@manager.be, pw:admin, region:antwerp)...')
 
-    if solr_install_path:
-        os.chdir(solr_install_path)
-        temp_command = '.\\bin\\solr' if os.name == 'nt' else './bin/solr'
-        print('stopping solr if it is running...')
-        call([temp_command, 'stop', '-all'], shell=True)
-        sleep(5)
-        print('  done')
-
-        print('generating haystack schema.xml...')
-        with open('example\solr\collection1\conf\schema.xml', mode='w') as f:
-            call_command('build_solr_schema', stdout=f)
-        print('  done')
-
-        print('start solr...')
-        call([temp_command, 'start'], shell=True)
-        sleep(5)
-        print('  done')
-
-        print('rebuilding index...')
-        call_command('rebuild_index', '--noinput')
-        print('  done')
 
 
 if __name__ == '__main__':
